Sort items in the left bar by weight.
Permalinks defined in pages in the group basic-documentation put them in
the root of the web site instead of the documentation/startup/
directory. The permalinks were used since the file names had to contain
numbers to ensure the correct ordering in the left navigation bar.

This patch adds a weighted_pages_list macro in _includes that uses the
weight attribute defined in the page front matter for sorting. Note that
apage that does not define the weight attribute will not appear in the
output of weighted_pages_list.